Started integrating parts into the OS - removed OS test project
Put the existing IO Expander/UART MIDI code into a new project that we can put new parts into as they're done. Seems to work so far, but nothing is optimised so it's pretty jumpy. I think the encoders aren't being read quick enough so they often start reading turns in the wrong direction.

Tested this code with axoloti and successfully controlled a dial in a patch with encoder 1.